+++ app/Product/Controller/ProductController.php
@@ -3,8 +3,8 @@
 namespace App\Product\Controller;
 
 use App\Http\Controllers\BaseController;
+use App\Product\DTO\ProductDto;
 use Illuminate\Http\Request;
-use App\Models\Product;
 use App\Product\Services\ProductServiceImpl;
 
 class ProductController extends BaseController
@@ -20,8 +20,29 @@ class ProductController extends BaseController
         return response()->json($prod);
     }
 
-    public function show(Request $request)
+    public function store(Request $request)
     {
-        return response()->json(Product::find($request->id));
+        $prod = new ProductDto();
+        return response()->json($this->service->create($prod->toProductModel($request)), 201);
+    }
+
+    public function getById(Request $request)
+    {
+        return response()->json($this->service->getById($request->id));
+    }
+
+    public function update(Request $request, int $id)
+    {
+        $prod = $this->service->getById($id);
+        $prod = new ProductDto();
+        $res = $this->service->update($id, $prod->toProductModel($request));
+
+        return response()->json($res);
+    }
+
+    public function destroy(int $id)
+    {
+        $this->service->delete($id);
+        return response()->json(null, 204);
     }
 }
