+++ routes/web.php
@@ -3,7 +3,8 @@
 use Illuminate\Support\Facades\Route;
 use App\Product\Controller\ProductController;
 
-// Route::group(['prefix' => 'products'], function () {
-//     Route::get('/', [ProductController::class, 'index'])->name('products.index');
-// });
-Route::get('/products', [ProductController::class, 'index'])->name('products.index');
+Route::group(['prefix' => 'products'], function () {
+    Route::post('/store', [ProductController::class, 'store'])->name('products.store'); 
+    Route::get('/all', [ProductController::class, 'index'])->name('products.index');
+    Route::get('/{id}', [ProductController::class, 'getById'])->name('products.getById');
+});
